+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Raytracing/HDRaytracingAmbientOcclusion.RenderGraph.cs
@@ -5,6 +5,12 @@ namespace UnityEngine.Rendering.HighDefinition
 {
     partial class HDRaytracingAmbientOcclusion
     {
+        struct TraceAmbientOcclusionResult
+        {
+            public TextureHandle signalBuffer;
+            public TextureHandle velocityBuffer;
+        }
+
         public TextureHandle RenderRTAO(RenderGraph renderGraph, HDCamera hdCamera, TextureHandle depthPyramid, TextureHandle normalBuffer, TextureHandle motionVectors, TextureHandle rayCountTexture, ShaderVariablesRaytracing shaderVariablesRaytracing)
         {
             var settings = hdCamera.volumeStack.GetComponent<AmbientOcclusion>();
@@ -15,10 +21,10 @@ namespace UnityEngine.Rendering.HighDefinition
             {
                 // Trace the signal
                 AmbientOcclusionTraceParameters aoParameters = PrepareAmbientOcclusionTraceParameters(hdCamera, shaderVariablesRaytracing);
-                result = TraceAO(renderGraph, aoParameters, depthPyramid, normalBuffer, rayCountTexture);
+                TraceAmbientOcclusionResult traceResult = TraceAO(renderGraph, aoParameters, depthPyramid, normalBuffer, rayCountTexture);
 
                 // Denoise if required
-                result = DenoiseAO(renderGraph, hdCamera, result, depthPyramid, normalBuffer, motionVectors);
+                result = DenoiseAO(renderGraph, hdCamera, traceResult, depthPyramid, normalBuffer, motionVectors);
 
                 // Compose the result to be done
                 AmbientOcclusionComposeParameters aoComposeParameters = PrepareAmbientOcclusionComposeParameters(hdCamera, shaderVariablesRaytracing);
@@ -38,12 +44,15 @@ namespace UnityEngine.Rendering.HighDefinition
             public TextureHandle normalBuffer;
             public TextureHandle rayCountTexture;
             public TextureHandle outputTexture;
+            public TextureHandle velocityBuffer;
         }
 
-        TextureHandle TraceAO(RenderGraph renderGraph, in AmbientOcclusionTraceParameters parameters, TextureHandle depthPyramid, TextureHandle normalBuffer, TextureHandle rayCountTexture)
+        TraceAmbientOcclusionResult TraceAO(RenderGraph renderGraph, in AmbientOcclusionTraceParameters parameters, TextureHandle depthPyramid, TextureHandle normalBuffer, TextureHandle rayCountTexture)
         {
             using (var builder = renderGraph.AddRenderPass<TraceRTAOPassData>("Tracing the rays for RTAO", out var passData, ProfilingSampler.Get(HDProfileId.RaytracingAmbientOcclusion)))
             {
+                TraceAmbientOcclusionResult traceOutput;
+
                 builder.EnableAsyncCompute(false);
 
                 passData.parameters = parameters;
@@ -53,6 +62,8 @@ namespace UnityEngine.Rendering.HighDefinition
                 // Depending of if we will have to denoise (or not), we need to allocate the final format, or a bigger texture
                 passData.outputTexture = builder.WriteTexture(renderGraph.CreateTexture(new TextureDesc(Vector2.one, true, true)
                     { colorFormat = GraphicsFormat.R8_UNorm, enableRandomWrite = true, name = "Ray Traced Ambient Occlusion" }));
+                passData.velocityBuffer = builder.ReadWriteTexture(renderGraph.CreateTexture(new TextureDesc(Vector2.one, true, true)
+                    { colorFormat = GraphicsFormat.R8_SNorm, enableRandomWrite = true, name = "Velocity Buffer" }));
 
                 builder.SetRenderFunc(
                     (TraceRTAOPassData data, RenderGraphContext ctx) =>
@@ -63,15 +74,18 @@ namespace UnityEngine.Rendering.HighDefinition
                         aotResources.normalBuffer = data.normalBuffer;
                         aotResources.rayCountTexture = data.rayCountTexture;
                         aotResources.outputTexture = data.outputTexture;
+                        aotResources.velocityBuffer = data.velocityBuffer;
 
                         TraceAO(ctx.cmd, data.parameters, aotResources);
                     });
 
-                return passData.outputTexture;
+                traceOutput.signalBuffer = passData.outputTexture;
+                traceOutput.velocityBuffer = passData.velocityBuffer;
+                return traceOutput;
             }
         }
 
-        TextureHandle DenoiseAO(RenderGraph renderGraph, HDCamera hdCamera, TextureHandle rayTracedAO, TextureHandle depthPyramid, TextureHandle normalBuffer, TextureHandle motionVectorBuffer)
+        TextureHandle DenoiseAO(RenderGraph renderGraph, HDCamera hdCamera, TraceAmbientOcclusionResult traceAOResult, TextureHandle depthPyramid, TextureHandle normalBuffer, TextureHandle motionVectorBuffer)
         {
             var aoSettings = hdCamera.volumeStack.GetComponent<AmbientOcclusion>();
             if (aoSettings.denoise)
@@ -83,17 +97,17 @@ namespace UnityEngine.Rendering.HighDefinition
                 HDTemporalFilter temporalFilter = m_RenderPipeline.GetTemporalFilter();
                 TemporalFilterParameters tfParameters = temporalFilter.PrepareTemporalFilterParameters(hdCamera, true, historyValidity);
                 TextureHandle historyBuffer = renderGraph.ImportTexture(RequestAmbientOcclusionHistoryTexture(hdCamera));
-                TextureHandle denoisedRTAO = temporalFilter.Denoise(renderGraph, hdCamera, tfParameters, rayTracedAO, historyBuffer, depthPyramid, normalBuffer, motionVectorBuffer);
+                TextureHandle denoisedRTAO = temporalFilter.Denoise(renderGraph, hdCamera, tfParameters, traceAOResult.signalBuffer, traceAOResult.velocityBuffer, historyBuffer, depthPyramid, normalBuffer, motionVectorBuffer);
 
                 // Apply the diffuse denoiser
                 HDDiffuseDenoiser diffuseDenoiser = m_RenderPipeline.GetDiffuseDenoiser();
                 DiffuseDenoiserParameters ddParams = diffuseDenoiser.PrepareDiffuseDenoiserParameters(hdCamera, true, aoSettings.denoiserRadius, false, false);
-                rayTracedAO = diffuseDenoiser.Denoise(renderGraph, hdCamera, ddParams, denoisedRTAO, depthPyramid, normalBuffer, rayTracedAO);
+                traceAOResult.signalBuffer = diffuseDenoiser.Denoise(renderGraph, hdCamera, ddParams, denoisedRTAO, depthPyramid, normalBuffer, traceAOResult.signalBuffer);
 
-                return rayTracedAO;
+                return traceAOResult.signalBuffer;
             }
             else
-                return rayTracedAO;
+                return traceAOResult.signalBuffer;
         }
 
         class ComposeRTAOPassData